+++ xdef/changelog.md
@@ -1,6 +1,6 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 42.2.22, release-date 2025-03-06
+# Version 42.2.22, release-date 2025-03-07
 * Fixed an error when using the `xdef_specdates` property (see `org.xdef.XDConstants.XDPROPERTY_SPECDATES`)
   when the date list contains a date without a time specification and the parsed dateTime has a time
   specified as 00:00:00.